requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.121.3
feedparser==6.0.12
filelock==3.20.0
Flask==3.1.1
fonttools==4.59.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.7.0
narwhals==2.12.0
networkx==3.5
numpy==2.3.3
openai==2.8.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
postgrest==2.24.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
realtime==2.24.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
storage3==2.24.0
streamlit==1.51.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
